+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Czech translation (Thanks to @office256) 
 - TimeSheet description and tags are saved on stop [#43](https://github.com/AlexandrePTJ/kemai/issues/43).
+- Option to enable/disable Kemai's update check [#68](https://github.com/AlexandrePTJ/kemai/issues/68).
+- Can stop current timesheet when idle is detected [#63](https://github.com/AlexandrePTJ/kemai/issues/63).
 
 ### Fixed
 - Enhance MacOSX generated dmg [#54](https://github.com/AlexandrePTJ/kemai/issues/54).